Reorganize components
Create shared 'progress-indicators' folder.
Move CircleTimer, FillLoader, FullScreenLoader, and WalletSyncCircle into this folder.
