+++ src/components/progress-indicators/WalletSyncCircle.js
