+++ src/components/scenes/CryptoExchangeQuoteScene.js
@@ -14,10 +14,10 @@ import { type RouteProp } from '../../types/routerTypes.js'
 import { type GuiSwapInfo } from '../../types/types.js'
 import { getSwapPluginIconUri } from '../../util/CdnUris'
 import { logEvent } from '../../util/tracking.js'
-import { CircleTimer } from '../common/CircleTimer'
 import { SceneWrapper } from '../common/SceneWrapper.js'
 import { ButtonsModal } from '../modals/ButtonsModal'
 import { swapVerifyTerms } from '../modals/SwapVerifyTermsModal.js'
+import { CircleTimer } from '../progress-indicators/CircleTimer'
 import { Airship, showError } from '../services/AirshipInstance'
 import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../services/ThemeContext.js'
 import { Alert } from '../themed/Alert'
