+++ src/components/scenes/FioRequestListScene.js
@@ -15,11 +15,11 @@ import { getExchangeDenominationFromState } from '../../selectors/DenominationSe
 import { connect } from '../../types/reactRedux.js'
 import { type NavigationProp } from '../../types/routerTypes.js'
 import type { FioAddress, FioRequest, GuiWallet } from '../../types/types'
-import { FullScreenLoader } from '../common/FullScreenLoader'
 import { SceneWrapper } from '../common/SceneWrapper'
 import { ButtonsModal } from '../modals/ButtonsModal'
 import type { WalletListResult } from '../modals/WalletListModal'
 import { WalletListModal } from '../modals/WalletListModal'
+import { FullScreenLoader } from '../progress-indicators/FullScreenLoader'
 import { Airship, showError, showToast } from '../services/AirshipInstance.js'
 import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../services/ThemeContext'
 import { EdgeText } from '../themed/EdgeText'
