+++ src/components/scenes/Staking/StakeModifyScene.js
@@ -16,10 +16,10 @@ import type { NavigationProp, RouteProp } from '../../../types/routerTypes'
 import { getCurrencyIconUris } from '../../../util/CdnUris'
 import { getPolicyIconUris, getPolicyTitleName, getPositionAllocations, stakePlugin } from '../../../util/stakeUtils.js'
 import { zeroString } from '../../../util/utils.js'
-import { FillLoader } from '../../common/FillLoader.js'
 import { SceneWrapper } from '../../common/SceneWrapper.js'
 import { FlipInputModal } from '../../modals/FlipInputModal.js'
 import { FlashNotification } from '../../navigation/FlashNotification.js'
+import { FillLoader } from '../../progress-indicators/FillLoader'
 import { Airship } from '../../services/AirshipInstance.js'
 import { cacheStyles, useTheme } from '../../services/ThemeContext.js'
 import { Alert } from '../../themed/Alert.js'
