+++ src/components/scenes/Staking/StakeOverviewScene.js
@@ -16,8 +16,8 @@ import { type NavigationProp } from '../../../types/routerTypes.js'
 import { guessFromCurrencyCode } from '../../../util/CurrencyInfoHelpers'
 import { getAllocationLocktimeMessage, getPolicyIconUris, getPolicyTitleName, getPositionAllocations, stakePlugin } from '../../../util/stakeUtils.js'
 import { StakingReturnsCard } from '../../cards/StakingReturnsCard.js'
-import { FillLoader } from '../../common/FillLoader'
 import { SceneWrapper } from '../../common/SceneWrapper.js'
+import { FillLoader } from '../../progress-indicators/FillLoader'
 import { cacheStyles, useTheme } from '../../services/ThemeContext.js'
 import { MainButton } from '../../themed/MainButton.js'
 import { SceneHeader } from '../../themed/SceneHeader.js'
